refactor: Move Launchpad tests into tests/integration/
The Launchpad crash DB tests should not be inside the `apport` module to
avoid installing them on all systems. Therefore move it to the
integration tests.

Previously the Launchpad tests could be run by calling:

```
TEST_LAUNCHPAD=1 python3 apport/crashdb_impl/launchpad.py
```

Now they can be run by calling:

```
TEST_LAUNCHPAD=1 python3 -m unittest tests/integration/test_crashdb_launchpad.py
```

The tests log in to https://qastaging.launchpad.net and the user need to
have bug control permissions on the Ubuntu coreutils package. They need
around seven minutes to execute.
